Move the ugly workaround code to utils.py, next attempt at solving the packaging unicode issue
